Use prod builds, add v3.28.8 and v4.0.1, add build instructions
Increasingly, Ember relies on the ember-cli defaults, and other packages. The new instructions in the README will ensure we get all the default packages/defaults included in the bundle, and should make it easier to add new versions in the future.

The new instructions also give us a true 'production' build. The existing builds we're using have been taken directly from the `ember-source` package, and include debug-only logic which may affect performance.

This commit replaces all existing versions with ember-cli production builds, as well as introducing 3.8.3 and 4.0.1